Pull offline imagery logic to AbstractMapContainerFragment
This allows showing the offline imagery in tasks and any fragment extending the base fragment
